Add "include..." special note

+++ jupyter_c_kernel/kernel.py
@@ -98,7 +98,7 @@ class CKernel(Kernel):
 
     main_head = "#include <stdio.h>\n" \
             "#include <math.h>\n" \
-            "int main(){\n"
+            "int main(int argc, char* argv[], char** env){\n"
 
     main_foot = "\nreturn 0;\n}"
 
@@ -145,6 +145,42 @@ class CKernel(Kernel):
     def _read_from_stdin(self):
         return self.raw_input()
 
+    def readcodefile(self,filename,spacecount=0):
+        filecode=''
+        codelist1=None
+        with open(os.path.join(os.path.abspath(''),filename), 'r') as codef1:
+            codelist1 = codef1.readlines()
+        if len(codelist1)>0:
+            for t in codelist1:
+                filecode+=' '*spacecount + t
+        return filecode
+   
+    def do_shell_command(self,commands=None,cwd=None):
+        p = self.create_jupyter_subprocess(['dart']+commands,cwd=os.path.abspath(''),shell=False)
+        while p.poll() is None:
+            p.write_contents()
+        # wait for threads to finish, so output is always shown
+        p._stdout_thread.join()
+        p._stderr_thread.join()
+
+        p.write_contents()
+
+        if p.returncode != 0:
+            self._write_to_stderr("[Dart kernel] Executable exited with code {}".format(p.returncode))
+        else:
+            self._write_to_stdout("[Dart kernel] Info:dart command success.")
+        return
+    
+    def create_jupyter_subprocess(self, cmd,cwd=None,shell=False):
+        return RealTimeSubprocess(cmd,
+                                  self._write_to_stdout,
+                                  self._write_to_stderr,
+                                  self._read_from_stdin,cwd,shell)
+    
+    def generate_file(self, source_filename, binary_filename, cflags=None, ldflags=None):
+
+        return
+    
     def create_jupyter_subprocess(self, cmd):
         return RealTimeSubprocess(cmd,
                                   self._write_to_stdout,
@@ -157,6 +193,7 @@ class CKernel(Kernel):
         # cflags = ['-std=iso9899:199409', '-pedantic', '-fPIC', '-shared', '-rdynamic'] + cflags
         # cflags = ['-std=c99', '-pedantic', '-fPIC', '-shared', '-rdynamic'] + cflags
         # cflags = ['-std=c11', '-pedantic', '-fPIC', '-shared', '-rdynamic'] + cflags
+        outfile=''
         if self.linkMaths:
             cflags = cflags + ['-lm']
         if self.wError:
@@ -167,25 +204,55 @@ class CKernel(Kernel):
             cflags = ['-DREAD_ONLY_FILE_SYSTEM'] + cflags
         if self.bufferedOutput:
             cflags = ['-DBUFFERED_OUTPUT'] + cflags
+        if '-o' not in cflags:
+            outfile=binary_filename
+        else:
+            outfile=cflags[cflags.index('-o')+1]
+            binary_filename=outfile
         args = ['gcc', source_filename] + cflags + ['-o', binary_filename] + ldflags
-        return self.create_jupyter_subprocess(args)
+        # self._write_to_stdout(''.join((' '+ str(s) for s in args)))
+        return self.create_jupyter_subprocess(args),binary_filename
 
     def _filter_magics(self, code):
 
         magics = {'cflags': [],
                   'ldflags': [],
+                  'file': [],
+                  'norun': [],
+                  'include': [],
+                  'command': [],
                   'args': []}
 
         actualCode = ''
 
         for line in code.splitlines():
-            if line.startswith('//%'):
-                key, value = line[3:].split(":", 2)
+            orgline=line
+            if line.strip().startswith('//%'):
+                if line.strip()[3:] == "norun":
+                    magics['norun'] += ['true']
+                    continue
+
+                key, value = line.strip()[3:].split(":", 2)
                 key = key.strip().lower()
 
                 if key in ['ldflags', 'cflags']:
                     for flag in value.split():
                         magics[key] += [flag]
+                elif key == "file":
+                    for flag in value.split():
+                        magics[key] += [flag]
+                elif key == "include":
+                    for flag in value.split():
+                        magics[key] += [flag]
+                    if len(magics['include'])>0:
+                        index1=line.find('//%')
+                        line=self.readcodefile(magics['include'][0],index1)
+                        actualCode += line + '\n'
+                elif key == "command":
+                    for flag in value.split():
+                        magics[key] += [flag]
+                    if len(magics['command'])>0:
+                        self.do_dart_command(magics['command'])
                 elif key == "args":
                     # Split arguments respecting quotes
                     for argument in re.findall(r'(?:[^\s,"]|"(?:\\.|[^"])*")+', value):
@@ -199,7 +266,6 @@ class CKernel(Kernel):
                 actualCode += line + '\n'
 
         return magics, actualCode
-
     # check whether int main() is specified, if not add it around the code
     # also add common magics like -lm
     def _add_main(self, magics, code):
@@ -207,7 +273,7 @@ class CKernel(Kernel):
         tmpCode = re.sub(r"//.*", "", code)
         tmpCode = re.sub(r"/\*.*?\*/", "", tmpCode, flags=re.M|re.S)
 
-        x = re.search(r"int\s+main\s*\(", tmpCode)
+        x = re.search(r".*\s+main\s*\(", tmpCode)
 
         if not x:
             code = self.main_head + code + self.main_foot
@@ -217,36 +283,47 @@ class CKernel(Kernel):
 
     def do_execute(self, code, silent, store_history=True,
                    user_expressions=None, allow_stdin=True):
-
+ 
         magics, code = self._filter_magics(code)
+        if len(magics['norun'])<1:
+            magics, code = self._add_main(magics, code)
 
-        magics, code = self._add_main(magics, code)
+        # magics, code = self._add_main(magics, code)
 
         # replace stdio with wrapped version
-        headerDir = "\"" + self.resDir + "/stdio_wrap.h" + "\""
-        code = code.replace("<stdio.h>", headerDir)
-        code = code.replace("\"stdio.h\"", headerDir)
+        # headerDir = "\"" + self.resDir + "/stdio_wrap.h" + "\""
+        # code = code.replace("<stdio.h>", headerDir)
+        # code = code.replace("\"stdio.h\"", headerDir)
 
         with self.new_temp_file(suffix='.c') as source_file:
             source_file.write(code)
             source_file.flush()
+
+            # if len(magics['norun'])>0:
+            if len(magics['file'])>0:
+                jfile = magics['file'][0]
+                # for x in jfile: self._write_to_stderr("file " + x + " ")
+                os.rename(source_file.name,os.path.join(os.path.abspath(''),jfile))
+                source_file.name=os.path.join(os.path.abspath(''),jfile)
+                self._write_to_stdout("[C kernel] Info:file created successfully\n")
+            if len(magics['norun'])>0:
+                if len(magics['file'])<1:
+                    self._write_to_stderr("[C kernel] Warning: no file name parameter\n")
+                return {'status': 'ok', 'execution_count': self.execution_count, 'payload': [], 'user_expressions': {}}
             with self.new_temp_file(suffix='.out') as binary_file:
-                p = self.compile_with_gcc(source_file.name, binary_file.name, magics['cflags'], magics['ldflags'])
+                p,outfile = self.compile_with_gcc(source_file.name, binary_file.name, magics['cflags'], magics['ldflags'])
                 while p.poll() is None:
                     p.write_contents()
                 p.write_contents()
+                binary_file.name=os.path.join(os.path.abspath(''),outfile)
                 if p.returncode != 0:  # Compilation failed
-                    self._write_to_stderr(
-                            "[C kernel] GCC exited with code {}, the executable will not be executed".format(
-                                    p.returncode))
+                    self._write_to_stderr("[C kernel] GCC exited with code {}, the executable will not be executed".format(p.returncode))
 
                     # delete source files before exit
                     os.remove(source_file.name)
                     os.remove(binary_file.name)
 
-                    return {'status': 'ok', 'execution_count': self.execution_count, 'payload': [],
-                            'user_expressions': {}}
-
+                    return {'status': 'ok', 'execution_count': self.execution_count, 'payload': [],'user_expressions': {}}
         # p = self.create_jupyter_subprocess([self.master_path, binary_file.name] + magics['args'])
         p = self.create_jupyter_subprocess([binary_file.name] + magics['args'])
         while p.poll() is None:
@@ -259,8 +336,9 @@ class CKernel(Kernel):
         p.write_contents()
 
         # now remove the files we have just created
-        os.remove(source_file.name)
-        os.remove(binary_file.name)
+        if len(magics['file'])<1:
+            os.remove(source_file.name)
+            os.remove(binary_file.name)
 
         if p.returncode != 0:
             self._write_to_stderr("[C kernel] Executable exited with code {}".format(p.returncode))
